XC-497: move SimpleHeaderStore to fixtures
